requirements.txt
vllm>=0.4.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
runpod>=1.6.0
pyyaml>=6.0
httpx>=0.27.0
accelerate>=0.27.0
huggingface_hub>=0.21.0
requests>=2.31.0
python-dotenv>=1.0.0
